Decorators
This contains the definition of the decorators that may be used on a
processor function.  The @want() decorator may be used to select which
events the processor function should be called for; the @requires()
decorator may be used to specify other processors that should run first;
and the @required_by() decorator may be used to specify that the
decorated processor should run before the listed processors.