Eliminate extraneous requirement
This change eliminates an extraneous requirement on pbr, which was
unnecessarily restricted.  This caused a problem for another project
that used another package that needed a more recent version of pbr.

This change also ignores the E731 pep8 error, a new rule that
conflicts with the use of a lambda expression.  The alternative was
to turn the lambda expression into a formal function definition, but
that seemed excessive and unnecessary for this case.